# Convert LazyList to Seq