--- a/Hedgehog.sln
+++ b/Hedgehog.sln
@@ -1,6 +1,6 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27130.2024
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.28527.54
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution items", "Solution items", "{76532829-C6DB-48F1-ABA8-835E3B718023}"
 	ProjectSection(SolutionItems) = preProject
@@ -8,11 +8,13 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "Hedgehog", "src/Hedgehog/Hedgehog.fsproj", "{0864FFED-D0B3-4D32-BD38-DB696B7FC5F3}"
+Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "Hedgehog", "src\Hedgehog\Hedgehog.fsproj", "{0864FFED-D0B3-4D32-BD38-DB696B7FC5F3}"
 EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Hedgehog.Tests", "tests/Hedgehog.Tests/Hedgehog.Tests.fsproj", "{D38AB54A-D5E8-4B50-927C-472015473AFC}"
+Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "Hedgehog.Tests", "tests\Hedgehog.Tests\Hedgehog.Tests.fsproj", "{D38AB54A-D5E8-4B50-927C-472015473AFC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Hedgehog.CSharp.Tests", "tests/Hedgehog.CSharp.Tests/Hedgehog.CSharp.Tests.csproj", "{952547C5-7622-45C8-BB3C-4D1BA2DB747D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Hedgehog.CSharp.Tests", "tests\Hedgehog.CSharp.Tests\Hedgehog.CSharp.Tests.csproj", "{952547C5-7622-45C8-BB3C-4D1BA2DB747D}"
+EndProject
+Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Hedgehog.Benchmarks", "tests\Hedgehog.Benchmarks\Hedgehog.Benchmarks.fsproj", "{7E2DE9D9-3831-46D7-B80F-B643E9D53FD3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -32,6 +34,10 @@ Global
 		{952547C5-7622-45C8-BB3C-4D1BA2DB747D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{952547C5-7622-45C8-BB3C-4D1BA2DB747D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{952547C5-7622-45C8-BB3C-4D1BA2DB747D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7E2DE9D9-3831-46D7-B80F-B643E9D53FD3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7E2DE9D9-3831-46D7-B80F-B643E9D53FD3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7E2DE9D9-3831-46D7-B80F-B643E9D53FD3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7E2DE9D9-3831-46D7-B80F-B643E9D53FD3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE