--- a/src/Hedgehog/Tree.fs
+++ b/src/Hedgehog/Tree.fs
@@ -1,12 +1,10 @@
 ﻿namespace Hedgehog
 
-open FSharpx.Collections
-
 /// A rose tree which represents a random generated outcome, and all the ways
 /// in which it can be made smaller.
-type Tree<'a> =
-    | Node of 'a * LazyList<Tree<'a>>
-
+type Tree<'a> = 
+    | Node of 'a * seq<Tree<'a>>
+    
 module Tree =
     /// The generated outcome.
     let outcome (Node (x, _) : Tree<'a>) : 'a =
@@ -16,22 +14,22 @@ module Tree =
     /// smallest to largest as if property still fails with the first shrink in
     /// the list then we will commit to that path and none of the others will
     /// be tried (i.e. there is no backtracking).
-    let shrinks (Node (_, xs) : Tree<'a>) : LazyList<Tree<'a>> =
+    let shrinks (Node (_, xs) : Tree<'a>) : seq<Tree<'a>> =
         xs
 
     /// Create a tree with a single outcome and no shrinks.
     let singleton (x : 'a) : Tree<'a> =
-        Node (x, LazyList.empty)
+        Node (x, Seq.empty)
 
     /// Map over a tree.
     let rec map (f : 'a -> 'b) (Node (x, xs) : Tree<'a>) : Tree<'b> =
-        Node (f x, LazyList.map (map f) xs)
+        Node (f x, Seq.map (map f) xs)
 
     let rec bind (Node (x, xs0) : Tree<'a>) (k : 'a -> Tree<'b>) : Tree<'b> =
         match k x with
         | Node (y, ys) ->
-            let xs = LazyList.map (fun m -> bind m k) xs0
-            Node (y, LazyList.append xs ys)
+            let xs = Seq.map (fun m -> bind m k) xs0
+            Node (y, Seq.append xs ys)
 
     let join (xss : Tree<Tree<'a>>) : Tree<'a> =
         bind xss id
@@ -41,23 +39,23 @@ module Tree =
     /// while still allowing the property to shrink to a minimal
     /// counterexample.
     let rec duplicate (Node (_, ys) as x : Tree<'a>) : Tree<Tree<'a>> =
-        Node (x, LazyList.map duplicate ys)
+        Node (x, Seq.map duplicate ys)
 
     /// Fold over a tree.
-    let rec fold (f : 'a -> 'x -> 'b) (g : LazyList<'b> -> 'x) (Node (x, xs) : Tree<'a>) : 'b =
+    let rec fold (f : 'a -> 'x -> 'b) (g : seq<'b> -> 'x) (Node (x, xs) : Tree<'a>) : 'b =
         f x (foldForest f g xs)
 
     /// Fold over a list of trees.
-    and foldForest (f : 'a -> 'x -> 'b) (g : LazyList<'b> -> 'x) (xs : LazyList<Tree<'a>>) : 'x =
-        LazyList.map (fold f g) xs |> g
+    and foldForest (f : 'a -> 'x -> 'b) (g : seq<'b> -> 'x) (xs : seq<Tree<'a>>) : 'x =
+        Seq.map (fold f g) xs |> g
 
     /// Build a tree from an unfolding function and a seed value.
-    let rec unfold (f : 'b -> 'a) (g : 'b -> LazyList<'b>) (x : 'b) : Tree<'a> =
+    let rec unfold (f : 'b -> 'a) (g : 'b -> seq<'b>) (x : 'b) : Tree<'a> =
         Node (f x, unfoldForest f g x)
 
     /// Build a list of trees from an unfolding function and a seed value.
-    and unfoldForest (f : 'b -> 'a) (g : 'b -> LazyList<'b>) (x : 'b) : LazyList<Tree<'a>> =
-        g x |> LazyList.map (unfold f g)
+    and unfoldForest (f : 'b -> 'a) (g : 'b -> seq<'b>) (x : 'b) : seq<Tree<'a>> =
+        g x |> Seq.map (unfold f g)
 
     /// Apply an additional unfolding function to an existing tree.
     ///
@@ -69,7 +67,7 @@ module Tree =
     ///
     /// Tree.unfold f (outcome oldTree)
     ///
-    let rec expand (f : 'a -> LazyList<'a>) (Node (x, xs) : Tree<'a>) : Tree<'a> =
+    let rec expand (f : 'a -> seq<'a>) (Node (x, xs) : Tree<'a>) : Tree<'a> =
         //
         // Ideally we could put the 'unfoldForest' nodes before the 'map expandTree'
         // nodes, so that we're culling from the top down and we would be able to
@@ -77,9 +75,9 @@ module Tree =
         //
         // We'd need some kind of tree transpose to do this properly.
         //
-        let ys = LazyList.map (expand f) xs
+        let ys = Seq.map (expand f) xs
         let zs = unfoldForest id f x
-        Node (x, LazyList.append ys zs)
+        Node (x, Seq.append ys zs)
 
     /// Recursively discard any shrinks whose outcome does not pass the predicate.
     /// Note that the root outcome can never be discarded.
@@ -87,6 +85,6 @@ module Tree =
         Node (x, filterForest f xs)
 
     /// Recursively discard any trees whose outcome does not pass the predicate.
-    and filterForest (f : 'a -> bool) (xs : LazyList<Tree<'a>>) : LazyList<Tree<'a>> =
-        LazyList.filter (f << outcome) xs
-        |> LazyList.map (filter f)
+    and filterForest (f : 'a -> bool) (xs : seq<Tree<'a>>) : seq<Tree<'a>> =
+        Seq.filter (f << outcome) xs
+        |> Seq.map (filter f)