--- a/src/Hedgehog/paket.references
+++ b/src/Hedgehog/paket.references
@@ -1,9 +1,3 @@
 group Build
 
-	File: Collections.fs
-	File: LazyList.fsi
-	File: LazyList.fs
-
 	FSharp.Core
-
-	System.Collections.Specialized