--- a/tests/Hedgehog.Benchmarks/paket.references
+++ b/tests/Hedgehog.Benchmarks/paket.references
@@ -1,0 +1,5 @@
+group Test
+
+	FSharp.Core
+
+	BenchmarkDotNet