* Converted use of `LazyList` to `Seq`.
* Added some simple benchmarks using Benchmark.NET.
* Add struct to some single-case DUs.

 There's definitely more work that can be done performance-wise! In particular string generation is slow. The biggest culprit seems to be nested binds but I haven't dug into this deeply yet (for some reason `dotnet.exe` is segfaulting when I try to run under debugger...)

Overall effect is slight but measurable, going to `Seq` is a win in some cases and a loss in others. Adding judicious `[<Struct>]` recovers more performance.

1st (baseline, before any changes):

|              Method |        Mean |      Error |     StdDev |
|-------------------- |------------:|-----------:|-----------:|
|             GenInts |    397.1 us |   6.458 us |   6.041 us |
|     GenAsciiStrings | 15,215.3 us | 173.352 us | 162.153 us |
| BigExampleFromTests |  4,231.5 us |  83.257 us | 164.341 us |

|      Method |    N |       Mean |     Error |    StdDev |
|------------ |----- |-----------:|----------:|----------:|
| ForLoopTest |  100 |   7.758 ms | 0.0564 ms | 0.0528 ms |
| ForLoopTest | 1000 | 124.197 ms | 1.0000 ms | 0.9354 ms |



2nd (converting LazyList->Seq):

|              Method |        Mean |      Error |     StdDev |
|-------------------- |------------:|-----------:|-----------:|
|             GenInts |    399.8 us |   6.116 us |   4.775 us |
|     GenAsciiStrings | 14,194.2 us | 125.778 us | 111.499 us |
| BigExampleFromTests |  3,924.8 us |  76.334 us |  74.971 us |

|      Method |    N |       Mean |     Error |    StdDev |
|------------ |----- |-----------:|----------:|----------:|
| ForLoopTest |  100 |   8.308 ms | 0.1848 ms | 0.2985 ms |
| ForLoopTest | 1000 | 129.990 ms | 1.4900 ms | 1.3209 ms |

3rd (after converting some things to structs):

|              Method |        Mean |      Error |     StdDev |
|-------------------- |------------:|-----------:|-----------:|
|             GenInts |    373.0 us |   4.142 us |   3.234 us |
|     GenAsciiStrings | 14,624.9 us | 187.147 us | 175.057 us |
| BigExampleFromTests |  3,657.5 us |  64.775 us |  60.591 us |

|      Method |    N |       Mean |     Error |    StdDev |
|------------ |----- |-----------:|----------:|----------:|
| ForLoopTest |  100 |   6.939 ms | 0.0743 ms | 0.0659 ms |
| ForLoopTest | 1000 | 116.753 ms | 1.9327 ms | 1.8078 ms |


Some more things I'd like to do:

- [ ] Add a wrapper around the `seq`s that make sure we never regress and enumerate them multiple times.
- [ ] Check memory allocations before/after.