@Porges 2019-02-05
I also fixed some build breaks with recent F# version. I'm not sure what's on the build server so let's see 😁