@Porges 2019-02-05
Definitely could reduce allocs more. Here's before/after adding struct annotations (ignore timings, laptop is doing a lot of other things):

### Before structs

|              Method |        Mean |      Error |     StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|-------------------- |------------:|-----------:|-----------:|------------:|------------:|------------:|--------------------:|
|             GenInts |    411.4 us |   7.884 us |   9.386 us |    204.1016 |           - |           - |           418.97 KB |
|     GenAsciiStrings | 15,772.6 us | 300.615 us | 334.133 us |   5875.0000 |           - |           - |         12085.22 KB |
| BigExampleFromTests |  4,156.9 us |  82.652 us | 121.150 us |   1492.1875 |     39.0625 |           - |          3159.67 KB |

|      Method |    N |       Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------ |----- |-----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
| ForLoopTest |  100 |   8.838 ms | 0.1729 ms | 0.2982 ms |   6031.2500 |           - |           - |            12.09 MB |
| ForLoopTest | 1000 | 139.539 ms | 3.1670 ms | 5.1141 ms |  23250.0000 |  10750.0000 |           - |           117.84 MB |

### After structs

|              Method |        Mean |      Error |    StdDev |      Median | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|-------------------- |------------:|-----------:|----------:|------------:|------------:|------------:|------------:|--------------------:|
|             GenInts |    390.7 us |   9.954 us |  28.72 us |    382.3 us |    182.6172 |           - |           - |           374.37 KB |
|     GenAsciiStrings | 15,842.4 us | 314.220 us | 612.86 us | 15,789.2 us |   5875.0000 |           - |           - |         12063.38 KB |
| BigExampleFromTests |  4,138.5 us |  83.279 us | 226.57 us |  4,063.3 us |   1367.1875 |     31.2500 |           - |          2858.41 KB |

|      Method |    N |       Mean |     Error |     StdDev |     Median | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------ |----- |-----------:|----------:|-----------:|-----------:|------------:|------------:|------------:|--------------------:|
| ForLoopTest |  100 |   8.299 ms | 0.2482 ms |  0.7122 ms |   8.098 ms |   4976.5625 |           - |           - |             9.96 MB |
| ForLoopTest | 1000 | 133.257 ms | 5.3970 ms | 15.3105 ms | 126.790 ms |  28000.0000 |   5666.6667 |           - |            97.16 MB |