@moodmosaic 2019-02-07
FWIW, I've started looking at this today (plus trying it on a couple of fairly-sized test suites).