@moodmosaic 2019-02-07
Just so that we have these in place. — On seq vs LazyList:

>`LazyList` computes each element only once regardless of how many times the list is traversed. In this way, it's closer to a sequence returned from `Seq.cache` (rather than a typical sequence). But, other than caching, `LazyList` behaves exactly like a list: it uses a list structure under the hood and supports pattern matching. So you might say: use `LazyList` instead of `seq` when you need list semantics and caching (in addition to laziness).
>
>Regarding both being infinite, `seq`'s memory usage is constant while `LazyList`'s is linear.
> -- https://stackoverflow.com/a/9201534/467754

From the LazyList docs:

>LazyLists are possibly-infinite, cached sequences. See also IEnumerable/Seq for uncached sequences. LazyLists normally involve delayed computations without side-effects. The results of these computations are cached and evaluations will be performed only once for each element of the lazy list. In contrast, for sequences (IEnumerable) recomputation happens each time an enumerator is created and the sequence traversed.
>
>LazyLists can represent cached, potentially-infinite computations. Because they are cached they may cause memory leaks if some active code or data structure maintains a live reference to the head of an infinite or very large lazy list while iterating it, or if a reference is maintained after the list is no longer required.
>
>Lazy lists may be matched using the LazyList.Cons and LazyList.Nil active patterns. These may force the computation of elements of the list.
> -- https://fsprojects.github.io/FSharpx.Collections/reference/fsharpx-collections-lazylist-1.html

I think we could have very well used `seq` form the beginning.